+++ packages/core/src/lib/filter/index.ts
@@ -1,7 +1,4 @@
-import type { Stream } from "@libp2p/interface-connection";
 import type { Libp2p } from "@libp2p/interface-libp2p";
-import type { PeerId } from "@libp2p/interface-peer-id";
-import type { PeerStore } from "@libp2p/interface-peer-store";
 import type { Peer } from "@libp2p/interface-peer-store";
 import type { IncomingStreamData } from "@libp2p/interface-registrar";
 import type {
@@ -14,17 +11,12 @@ import type {
   ProtocolOptions,
 } from "@waku/interfaces";
 import { WakuMessage as WakuMessageProto } from "@waku/proto";
-import {
-  getPeersForProtocol,
-  selectConnection,
-  selectPeerForProtocol,
-  selectRandomPeer,
-} from "@waku/utils";
 import debug from "debug";
 import all from "it-all";
 import * as lp from "it-length-prefixed";
 import { pipe } from "it-pipe";
 
+import { BaseProtocol } from "../base_protocol.js";
 import { DefaultPubSubTopic } from "../constants.js";
 import { groupByContentTopic } from "../group_by.js";
 import { toProtoMessage } from "../to_proto_message.js";
@@ -46,8 +38,7 @@ export type UnsubscribeFunction = () => Promise<void>;
  * - https://github.com/status-im/go-waku/issues/245
  * - https://github.com/status-im/nwaku/issues/948
  */
-class Filter implements IFilter {
-  multicodec: string;
+class Filter extends BaseProtocol implements IFilter {
   options: ProtocolCreateOptions;
   private subscriptions: Map<string, Callback<any>>;
   private decoders: Map<
@@ -56,8 +47,8 @@ class Filter implements IFilter {
   >;
 
   constructor(public libp2p: Libp2p, options?: ProtocolCreateOptions) {
+    super(FilterCodec, libp2p);
     this.options = options ?? {};
-    this.multicodec = FilterCodec;
     this.subscriptions = new Map();
     this.decoders = new Map();
     this.libp2p
@@ -130,10 +121,6 @@ class Filter implements IFilter {
       this.deleteDecoders(groupedDecoders);
       this.deleteCallback(requestId);
     };
-  }
-
-  get peerStore(): PeerStore {
-    return this.libp2p.peerStore;
   }
 
   private onRequest(streamData: IncomingStreamData): void {
@@ -255,33 +242,6 @@ class Filter implements IFilter {
       log("Error unsubscribing", e);
       throw e;
     }
-  }
-
-  private async newStream(peer: Peer): Promise<Stream> {
-    const connections = this.libp2p.getConnections(peer.id);
-    const connection = selectConnection(connections);
-    if (!connection) {
-      throw new Error("Failed to get a connection to the peer");
-    }
-
-    return connection.newStream(this.multicodec);
-  }
-
-  private async getPeer(peerId?: PeerId): Promise<Peer> {
-    const { peer } = await selectPeerForProtocol(
-      this.peerStore,
-      [this.multicodec],
-      peerId
-    );
-    return peer;
-  }
-
-  async peers(): Promise<Peer[]> {
-    return getPeersForProtocol(this.peerStore, [this.multicodec]);
-  }
-
-  async randomPeer(): Promise<Peer | undefined> {
-    return selectRandomPeer(await this.peers());
   }
 }
