+++ packages/core/src/lib/light_push/index.ts
@@ -1,7 +1,5 @@
 import type { Libp2p } from "@libp2p/interface-libp2p";
 import type { PeerId } from "@libp2p/interface-peer-id";
-import type { Peer } from "@libp2p/interface-peer-store";
-import type { PeerStore } from "@libp2p/interface-peer-store";
 import type {
   IEncoder,
   ILightPush,
@@ -11,18 +9,13 @@ import type {
   SendResult,
 } from "@waku/interfaces";
 import { PushResponse } from "@waku/proto";
-import {
-  getPeersForProtocol,
-  selectConnection,
-  selectPeerForProtocol,
-  selectRandomPeer,
-} from "@waku/utils";
 import debug from "debug";
 import all from "it-all";
 import * as lp from "it-length-prefixed";
 import { pipe } from "it-pipe";
 import { Uint8ArrayList } from "uint8arraylist";
 
+import { BaseProtocol } from "../base_protocol.js";
 import { DefaultPubSubTopic } from "../constants.js";
 
 import { PushRPC } from "./push_rpc.js";
@@ -35,12 +28,11 @@ export { PushResponse };
 /**
  * Implements the [Waku v2 Light Push protocol](https://rfc.vac.dev/spec/19/).
  */
-class LightPush implements ILightPush {
-  multicodec: string;
+class LightPush extends BaseProtocol implements ILightPush {
   options: ProtocolCreateOptions;
 
   constructor(public libp2p: Libp2p, options?: ProtocolCreateOptions) {
-    this.multicodec = LightPushCodec;
+    super(LightPushCodec, libp2p);
     this.options = options || {};
   }
 
@@ -51,23 +43,8 @@ class LightPush implements ILightPush {
   ): Promise<SendResult> {
     const { pubSubTopic = DefaultPubSubTopic } = this.options;
 
-    const res = await selectPeerForProtocol(
-      this.peerStore,
-      [this.multicodec],
-      opts?.peerId
-    );
-
-    if (!res) {
-      throw new Error("Failed to get a peer");
-    }
-    const { peer } = res;
-
-    const connections = this.libp2p.getConnections(peer.id);
-    const connection = selectConnection(connections);
-
-    if (!connection) throw "Failed to get a connection to the peer";
-
-    const stream = await connection.newStream(this.multicodec);
+    const peer = await this.getPeer(opts?.peerId);
+    const stream = await this.newStream(peer);
 
     const recipients: PeerId[] = [];
 
@@ -108,28 +85,6 @@ class LightPush implements ILightPush {
       log("Failed to send waku light push request", err);
     }
     return { recipients };
-  }
-
-  /**
-   * Returns known peers from the address book (`libp2p.peerStore`) that support
-   * light push protocol. Waku may or may not be currently connected to these
-   * peers.
-   */
-  async peers(): Promise<Peer[]> {
-    return getPeersForProtocol(this.peerStore, [this.multicodec]);
-  }
-
-  /**
-   * Returns a random peer that supports light push protocol from the address
-   * book (`libp2p.peerStore`). Waku may or  may not be currently connected to
-   * this peer.
-   */
-  async randomPeer(): Promise<Peer | undefined> {
-    return selectRandomPeer(await this.peers());
-  }
-
-  get peerStore(): PeerStore {
-    return this.libp2p.peerStore;
   }
 }
